Add a gitlab-ci job to push images to dockerhub
It is configured to run only on the master branch (for now) and only if
the environment variable DOCKERHUB_IMAGE is set in the CI/CD settings.
In order for it to work, DOCKERHUB_USER and DOCKERHUB_PASSWORD variables
must be also present, they should be for a dockerhub service user
account.