Added support for multiple shared secrets.
In environments where multiple applications are integrated with a single
scalelite server, it might be sensible to give each application its own
secret. If an application secret is compromized, it can be revoked
without disturbing any of the other applications.

This patch adds an additional `LOADBALANCER_SECRETS` config knob that may
contain additional secrets, separated by `:`. Any of the configured secrets
will work. The change is fully backwards compatible.