Run task status in Parallel to connect to servers
If we have a server is offline, the task status currently waits for TIMEOUT (10 seconds) to connect in the next server.
If there are a large number of servers offline, it's can take a long time!

These changes I made are to parallelize the execution using threads.

Sorry if it wasn't a good job. I know very little about ruby.