SL-83: Make timeouts on requests to BBB api configurable (#440)
* SL-83: Make timeouts on requests to BBB api configurable

This is part of the work to solve #439. Simply making the timeouts
shorter and allowing them to be configured to match the environment of a
particular deployment should reduce the impact of attempting to make
calls on unreachable/slow BigBlueButton servers

* SL-83: Add dynamic timeout shortening on offline servers

On servers that the poller has marked as offline, shorten the request
timeouts. We're not really expecting the request to succeed, and it's
important to not tie up a thread for too long on a request that we think
is going to fail.

If it does go through, tho, that's a nice bonus.