Allow using server hostname as ID instead of UUID (#470)
* Ensure server ids are unique

Now that we allow people to use arbitrary strings (e.g. server
hostnames) as ids, they're no longer guaranteed to be unique like the
uuids were. Add protection to ensure that an attempt to add a server
with an existing id will fail.

* Allow configuring scalelite to use server hostnames as server ids

Off by default for now; that default might change in a later major
version of Scalelite.

Co-authored-by: Jesus Federico <jesus@123it.ca>