GIT-262: Fix setting SCALELITE_API_PORT env var in nginx config
The assignment to the variable was incorrectly prefixed with $, so it
wasn't getting set correctly; and besides that the variable wasn't
exported to the environment so the 'envsubst' tool wouldn't pick it up.

Note that since the shell used is /bin/sh (which can be 'dash' in some
distros), the more compatible form of using export separately from
setting the variable is used here.

The way the envsubst tool was being run was incorrect; the second time
it was run it would replace the file without setting the api port at
all. Fix that by substituting both variables in a single call to
envsubst.