Fix create tests
A lot of the create tests didn't handle the addition of the voiceBridge
parameter, due to relying on matching a fixed url (the voice bridge
numbers are randomly generated).

Rework the tests to run assertions on the mocked http requests in the
response generation handler.

While I'm at it, also add some checks to see that the requests do
actually get made - this revealed that some of the callback tests
weren't actually testing the right code.