SL-244: Add voice bridge allocation & freeswitch dialplan generation. (closes #769) (#797)
* Add documentation for voice bridge storage in redis

* Meeting model work for voice bridge generation/storage.

* Fix issues in voice_bridge allocation causing test failures

* Rework create meeting to immediately return existing meeting.

Need to avoid running the conference number allocation if the meeting is
already present, and this saves some work.
But the requirements for doing this mean putting a watch on the
meetings set, so meeting creation can now return a
ConcurrentModificationError

* Have the BigBlueButton create api call generate and send voice bridge.

* Fix create tests

A lot of the create tests didn't handle the addition of the voiceBridge
parameter, due to relying on matching a fixed url (the voice bridge
numbers are randomly generated).

Rework the tests to run assertions on the mocked http requests in the
response generation handler.

While I'm at it, also add some checks to see that the requests do
actually get made - this revealed that some of the callback tests
weren't actually testing the right code.

* Meeting model tests for conference number allocation

* Move create tests from rspec to minitest

I'm too lazy to figure out how to properly handle the voice bridge stuff
in rspec, so it's easier for me to move these tests to minitest instead.

* Also port the join tests to minitest so rspec can be removed

Much more straightforwards to only have 1 test framework instead of 2.

* Fix test voice bridge regexp to do string start/end matching

* Simplify create tests by using webmock parameter matching where possible.

* Update rubocop-rails to improve DynamicFindBy, switch to released fakeredis

* Freeswitch dialplan api for bridging calls

* Disable csrf protection on fsapi

* Fix fsapi route to use POST

* Handle the breakout room retry also returning an exception.

* Update fsapi tests to use POST request

* Fix or mask new errors exposed by rubocop-rails update.

* Update readme, add dialin setup documentation.

Co-authored-by: Jesus Federico <jesus@123it.ca>