docs: note to install and enable mod_xml_curl for dial-in (#1214)
The call gateway setup for dial-in using FreeSWITCH needs mod_xml_curl to work. 

This is not explicit in the README - and the module isn't compiled nor installed by default in FreeSWITCH.

Add a reminder that it should be both installed and enabled for the call gateway setup to work properly.